added frontend functionality

+++ routes/ideas.js
@@ -68,19 +68,28 @@ router.post('/', async (req, res) => {
 // Update idea
 router.put('/:id', async (req, res) => { 
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate( 
-      req.params.id,
-      { 
-        $set: { 
-          text: req.body.text,
-          tag: req.body.tag
+    const idea = await Idea.findById(req.params.id)
+
+    // Match the usernames
+    if (idea.username === req.body.username) {  
+      const updatedIdea = await Idea.findByIdAndUpdate( 
+        req.params.id,
+        { 
+          $set: { 
+            text: req.body.text,
+            tag: req.body.tag
+          }
+        },
+        { 
+          new: true
         }
-      },
-      { 
-        new: true
-      }
-    )
-    res.json({ success: true, data: updatedIdea })
+      )
+      res.json({ success: true, data: updatedIdea })
+    }    
+
+    // Usernames do not match
+    res.status(403).json({ success: false, error: 'You are not authorized to update this resource' })
+
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' })
@@ -90,8 +99,17 @@ router.put('/:id', async (req, res) => {
 // Delete idea
 router.delete('/:id', async (req, res) => { 
   try {
-    await Idea.findByIdAndDelete(req.params.id)
-    res.json({ success: true, data: {} })
+    const idea = await Idea.findById(req.params.id)
+
+    // Match the usernames
+    if (idea.username === req.body.username) { 
+      await Idea.findByIdAndDelete(req.params.id)
+      return res.json({ success: true, data: {} })
+    }
+
+    // Usernames do not match
+    res.status(403).json({ success: false, error: 'You are not authorized to delete this resource' })
+    
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' })
